requirement.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docx2txt==0.9
emoji==2.15.0
executing==2.2.1
filelock==3.25.2
filetype==1.2.0
Flask==3.1.3
flask-cors==6.0.2
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
groq==1.1.2
h11==0.16.0
hf-xet==1.4.2
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.21
langchain-huggingface==1.2.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numba==0.64.0
numpy==2.4.3
olefile==0.47
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==3.0.1
parso==0.8.6
pinecone==8.1.0
pinecone-client==6.0.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.9.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.32
unstructured-client==0.42.10
urllib3==2.6.3
uuid_utils==0.14.1
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.7
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
pypdf
groq